+++ CHANGES.rst
@@ -1,3 +1,14 @@
+0.3.0
+-----
+* glfw.init and glfw.terminate now setup during `import glfw`
+* Updated package structure
+* Added tests, test fixtures and increased coverage
+* Updated travis build to use GLFW 3.2.1
+* Updated tox work flows
+* Updated examples with new package structure
+* Adds docker
+
+
 0.2.0
 -----